fix: адаптировать компоненты в layout
