+++ src/widgets/layout/index.ts
@@ -1,0 +1,1 @@
+export { Layout } from './layout';
