+++ src/widgets/main-wrapper/index.ts
@@ -1,0 +1,1 @@
+export { MainWrapper } from './main-wrapper';
