+++ src/widgets/page-layuot/index.ts
@@ -1,0 +1,1 @@
+export { PageLayout } from './page-layuot';
